# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Nonogram API
  version: "0.8.3"
servers:
  - url: http://localhost:8080/
    description: Development server
paths:
...
  /puzzle/list:
    get:
      summary: Used to get a list of puzzles
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            example: 1
          description: The number of a page
        - in: query
          name: size
          schema:
            type: integer
            example: 10
          description: The quantity of items to be displayed on a page
      tags:
       - Puzzle
      responses:
        "200":
          description: Returns a page with the list of puzzles
          content:
            text/html:
              schema:
                $ref: "#/components/schemas/Puzzles"
  /puzzle/create:
    get:
      summary: Used to get a puzzle creation page
      tags:
        - Puzzle
      responses:
        "200":
          description: Returns a page that is the first puzzle creation stage
          content:
            text/html:
              example: <html></html>
    post:
      summary: Used to initiate a puzzle creation
      tags:
        - Puzzle
      parameters:
        - in: query
          name: name
          schema:
            type: string
            example: My New Puzzle
          description: The name of a new puzzle
        - in: query
          name: width
          schema:
            type: integer
            minimum: 10
            maximum: 40
            example: 10
          description: The width of a new puzzle
        - in: query
          name: height
          schema:
            type: integer
            minimum: 10
            maximum: 40
            example: 10
          description: The height of a new puzzle
        - in: query
          name: fill
          schema:
            type: string
            example: MANUALLY
          description: The fill mode of a new puzzle creation
        - in: query
          name: image
          schema:
            type: string
            format: binary
          description: The image for a puzzle generation when option generate by image is selected
      responses:
        "200":
          description: Returns next stage page regarding to the selected puzzle creation option
          content:
            text/html:
              example: <html>...</html>
  /puzzle/fill:
    post:
      summary: Used to fill new puzzle with cells when option MANUALLY is selected
      tags:
        - Puzzle
      parameters:
        - in: query
          name: coordinates
          schema:
            type: array
            items:
              type: string
              example: 1:1
      responses:
        "200":
          description: Returns the same page with error messages
          content:
            text/html:
              example: <html>...</html>
        "301":
          description: Redirects to /puzzle/list when the puzzle filled successfully
          content:
            text/html:
              example: <html>...</html>
  /puzzle/save/image:
    post:
      summary: Used to save the puzzle generated by image
      tags:
        - Puzzle
      parameters:
        - in: query
          name: puzzleId
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "301":
          description: Redirects to /puzzle/list when the puzzle filled successfully
          content:
            text/html:
              example: <html>...</html>
...
  /puzzle/{puzzle_id}/view:
    get:
      summary: Used to get info about the certain puzzle
      tags:
        - Puzzle
      parameters:
        - in: path
          required: true
          name: puzzle_id
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          description: Returns a page with info about the puzzle 
          content:
            text/html:
              example: <html>...</html>
  /puzzle/{puzzle_id}/play:
    get:
      summary: Used to get playing field for a certain puzzle
      tags:
        - Puzzle
      parameters:
        - in: path
          required: true
          name: puzzle_id
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          description: Returns a page with playing field for the puzzle
          content:
            text/html:
              example: <html>...</html>
  /puzzle/{puzzle_id}/delete:
    get:
      summary: Used to get deletion page for a certain puzzle
      tags:
        - Puzzle
      parameters:
        - in: path
          required: true
          name: puzzle_id
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "200":
          description: Returns a page with playing field for the puzzle
          content:
            text/html:
              example: <html>...</html>
    post:
      summary: Used to approve the puzzle deletion
      tags:
        - Puzzle
      parameters:
        - in: path
          required: true
          name: puzzle_id
          schema:
            type: integer
            format: int64
            example: 1
        - in: query
          name: confirmation
          schema:
            type: string
            example: Puzzle Name
      responses:
        "200":
          description: Returns the same page with error messages
          content:
            text/html:
              example: <html>...</html>
        "301":
          description: Redirects to /puzzle/list when the puzzle deleted successfully
          content:
            text/html:
              example: <html>...</html>
...
components:
  schemas:
...
    Puzzle:
      type: object
      required:
        - name
        - width
        - height
        - user
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name: 
          type: string
          example: My Puzzle
        width:
          type: integer
          minimum: 10
          maximum: 40
          example: 10
        height:
          type: integer
          minimum: 10
          maximum: 40
          example: 10
        visible:
          type: boolean
          example: true
        generated: 
          type: boolean
          example: true
        user:
          $ref: "#/components/schemas/User"
        cells:
          $ref: "#/components/schemas/Cells"
    Puzzles:
      type: array
      items:
        $ref: "#/components/schemas/Puzzle"
    User:
      type: object
      required:
        - nickname
        - email
        - password
        - passwordConfirmation
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nickname: 
          type: string
          example: Player1
        email: 
          type: string
          example: player@test.org
        password:
          type: string
          example: qweQWE123
        passwordConfirmation:
          type: string
          example: qweQWE123
        activated:
          type: boolean
          example: true
        score:
          type: integer
          example: 200
        role:
          type: string
          example: PLAYER
    Cell:
      type: object
      required:
        - x
        - y
      properties:
        x: 
          type: integer
          example: 1
        y:
          type: integer
          example: 1
    Cells:
      type: array
      items:
        $ref: "#/components/schemas/Cell"